Rename loginReducer -> login
