+++ services/frontend/src/store/index.js
@@ -1,12 +1,12 @@
 import { combineReducers } from 'redux';
 import { configureStore } from '@reduxjs/toolkit';
 import { devToolsEnhancer } from 'redux-devtools-extension';
-import { loginReducer } from './login-reducers';
+import { loginReducer } from './login';
 import tasks from './tasks';
 
 const rootReducer = combineReducers({
     tasks,
-    loginReducer,
+    login: loginReducer,
 });
 
 const store = configureStore({
